requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.1
requests==2.32.3
urllib3==1.26.18
chardet==5.2.0